political_analysis_requirements.txt
# Political Document Analysis System Requirements

# Core LlamaIndex packages
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-readers-web>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-graph-stores-neo4j>=0.1.0
llama-index-tools-mcp>=0.4.0
llama-index-agent-openai>=0.1.0

# Web crawling dependencies
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
playwright>=1.40.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.2.0

# Graph database
neo4j>=5.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
tqdm>=4.66.0
PyYAML>=6.0.1

# Development and testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
jupyter>=1.0.0

# Optional dependencies for advanced features
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0